Creation of the port constant and adjusments

+++ backend/controllers/product.controller.js
@@ -57,10 +57,16 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ success: false, error: "Product not found" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, data: "Product deleted" });
   } catch (error) {
-    res.status(404).json({ success: false, error: "Product not found" });
+    console.log("Error to delete a product: ", error.message);
+    res.status(500).json({ success: false, error: "Server error" });
   }
 };
